v3.4.0
-- New stuff

* Add New Connection Button for Clients block sources
* New Json view (auto)
* New Named Pipe / Local socket client
* Add TDS7 decoder
* Send Packets to existing Packets Analyser
* Send Packets to new Packets Analyser
* Add incremental names for tabs

-- Various

* Fix Mac OSX build
* Win package update
* Openssl linkage
* Travis build (draft)
* Refactoring
* Migrating to QSharedPointer packet management
* Fix Java Style issues
* Codacity badge
* Upgrading to C++14
* Fix deprecated warnings
* Adjust filterdialog window size

+++ extras/burp_plugin/src/burp/PayloadProcessor.java
@@ -147,7 +147,7 @@ public class PayloadProcessor implements IIntruderPayloadProcessor {
 				returnValue = new byte[returnedVals.size()];
 				
 				for (int i = 0; i < returnedVals.size(); i++) {
-					returnValue[i] = returnedVals.get(i).byteValue();
+					returnValue[i] = returnedVals.get(i);
 				}
 				
 			//	System.out.println("Data block read: ".concat(helpers.bytesToString(returnValue)));
